backend/pyproject.toml
[project]
name = "scratchcard-backend"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "aiohttp>=3.13.3",
    "aiosqlite>=0.22.1",
    "apscheduler>=3.11.2",
    "beautifulsoup4>=4.14.3",
    "fastapi>=0.135.1",
    "geopy>=2.4.1",
    "httpx>=0.28.1",
    "lxml>=6.0.2",
    "openpyxl>=3.1.5",
    "playwright>=1.58.0",
    "pydantic>=2.12.5",
    "python-dotenv>=1.2.2",
    "python-jose[cryptography]>=3.5.0",
    "sentry-sdk[fastapi]>=2.54.0",
    "sqlalchemy>=2.0.47",
    "tenacity>=9.1.4",
    "unidecode>=1.4.0",
    "uvicorn[standard]>=0.41.0",
]

[dependency-groups]
dev = [
    "httpx>=0.28.1",
    "pytest>=9.0.2",
]

[tool.pytest.ini_options]
pythonpath = ["."]
testpaths = ["tests"]
